+++ src/CHANGELOG.md
@@ -5,7 +5,7 @@
 - changed `getFolders()` method to get all directories, instead of showing only folders containing `.pug` and `dat.yaml`
 
 ### 6.08.12
-- wcomp must now be -custEl
+- wcomp must now be -custel
 
 ### 6.08.9
 - version bump